backend/requirements/base.txt
Django==4.2
djangorestframework==3.14
django-cors-headers==4.3.1
django-decouple==2.1
djangorestframework-simplejwt==5.3.1